# Faster title alignment