@rcomer 2024-05-25
I admit I haven't actually figure out _how_ the first change made the given example faster.