@rcomer 2024-05-26
I think this would benefit from more digging: the fact that there is no measurable time difference without _sharex_ suggests that this method ought to be fast and the real problem for _sharex_ is deeper.