@rcomer 2024-05-26
More digging completed https://github.com/matplotlib/matplotlib/issues/26150#issuecomment-2132256785