@rcomer 2024-05-28
I think CodeCov was telling me that we didn't have any tests with a child axes and an automatically place title, so I added one.