@rcomer 2024-05-28
> Does it actually speed things up?

If I take this example from #26150
`fig_many_sharex.savefig(stream, format='svg', bbox_inches='tight')`

With `main` I get
```
477 ms ± 13.3 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)
```
with this branch
```
224 ms ± 6.7 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)
```
That example has no titles though.  If I add a title to each subplot I get
`main`
```
511 ms ± 9.67 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)
```
Branch
```
275 ms ± 6.5 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)
```
If I add a title to each subplot _and_ move the ticklabels to the top of each subplot
`main`
```
707 ms ± 13.7 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)
```
Branch
```
464 ms ± 9.57 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)
```