@jklymak 2024-05-28
We have a benchmarking test suite.  I wonder if these are caught by it?  @QuLogic maintains that - I actually forget where it is...