@rcomer 2024-05-28
> We have a benchmarking test suite

This is news to me!