@QuLogic 2024-05-30
> > We have a benchmarking test suite
> 
> This is news to me!

We do; it is [here](https://github.com/matplotlib/mpl-bench). However, we don't test `bbox_inches='tight'` or `sharex`, and unfortunately, I haven't run the benchmark in a long while now.